tests/wpt/web-platform-tests/css/css-tables/parsing/table-layout-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module Test: parsing table-layout with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-table-layout">
<meta name="assert" content="table-layout supports only the grammar 'auto | fixed'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("table-layout", "none");
test_invalid_value("table-layout", "auto fixed");
</script>
</body>
</html>
